refactor: extract test helpers and update e2e tests
